+++ src/shared/experience/expSite.ts
@@ -7,6 +7,7 @@
 import fs from 'node:fs';
 import path from 'node:path';
 import { Org, SfError } from '@salesforce/core';
+import axios from 'axios';
 
 export type SiteMetadata = {
   bundleName: string;
@@ -92,6 +93,28 @@ export class ExperienceSite {
     }>('SELECT Id, Name, LastModifiedDate, UrlPathPrefix, Status FROM Network');
     const experienceSites: string[] = result.records.map((record) => record.Name);
     return experienceSites;
+  }
+
+  /**
+   * Esablish a valid token for this local development session
+   *
+   * @returns sid token for proxied site requests
+   */
+  public async setupAuth(): Promise<string> {
+    let sidToken = ''; // Default to guest user access only
+
+    // Use environment variable for now if users want to just have guest access only
+    if (process.env.SITE_GUEST_ACCESS !== 'true') {
+      try {
+        const networkId = await this.getNetworkId();
+        sidToken = await this.getNewSidToken(networkId);
+      } catch (e) {
+        // eslint-disable-next-line no-console
+        console.error('Failed to establish authentication for site', e);
+      }
+    }
+
+    return sidToken;
   }
 
   public async isUpdateAvailable(): Promise<boolean> {
@@ -224,6 +247,99 @@ export class ExperienceSite {
     }
 
     return resourcePath;
+  }
+
+  private async getNetworkId(): Promise<string> {
+    const conn = this.org.getConnection();
+    // Query the Network object for the network with the given site name
+    const result = await conn.query<{ Id: string }>(`SELECT Id FROM Network WHERE Name = '${this.siteDisplayName}'`);
+
+    const record = result.records[0];
+    if (record) {
+      let networkId = record.Id;
+      // Subtract the last three characters from the Network ID
+      networkId = networkId.substring(0, networkId.length - 3);
+      return networkId;
+    } else {
+      throw new Error(`NetworkId for site: '${this.siteDisplayName}' could not be found`);
+    }
+  }
+
+  private async getNewSidToken(networkId: string): Promise<string> {
+    // Get the connection and access token from the org
+    const conn = this.org.getConnection();
+    const orgId = this.org.getOrgId();
+
+    // Not sure if we need to do this
+    const orgIdMinus3 = orgId.substring(0, orgId.length - 3);
+    const accessToken = conn.accessToken;
+    const instanceUrl = conn.instanceUrl; // Org URL
+
+    // Make the GET request without following redirects
+    if (accessToken) {
+      // TODO should we try and refresh auth here?
+      // await conn.refreshAuth();
+
+      // Call out to the switcher servlet to establish a session
+      const switchUrl = `${instanceUrl}/servlet/networks/switch?networkId=${networkId}`;
+      const cookies = [`sid=${accessToken}`, `oid=${orgIdMinus3}`].join('; ').trim();
+      let response = await axios.get(switchUrl, {
+        headers: {
+          Cookie: cookies,
+        },
+        withCredentials: true,
+        maxRedirects: 0, // Prevent axios from following redirects
+        validateStatus: (status) => status >= 200 && status < 400, // Accept 3xx status codes
+      });
+
+      // Extract the Location callback header
+      const locationHeader = response.headers['location'] as string;
+      if (locationHeader) {
+        // Parse the URL to extract the 'sid' parameter
+        const urlObj = new URL(locationHeader);
+        const sid = urlObj.searchParams.get('sid') ?? '';
+        const cookies2 = ['__Secure-has-sid=1', `sid=${sid}`, `oid=${orgIdMinus3}`].join('; ').trim();
+
+        // Request the location header to establish our session with the servlet
+        response = await axios.get(urlObj.toString(), {
+          headers: {
+            Cookie: cookies2,
+          },
+          withCredentials: true,
+          maxRedirects: 0, // Prevent axios from following redirects
+          validateStatus: (status) => status >= 200 && status < 400, // Accept 3xx status codes
+        });
+        const setCookieHeader = response.headers['set-cookie'];
+        if (setCookieHeader) {
+          // Find the 'sid' cookie in the set-cookie header
+          const sidCookie = setCookieHeader.find((cookieStr: string) => cookieStr.startsWith('sid='));
+          if (sidCookie) {
+            // Extract the sid value from the set-cookie string
+            const sidMatch = sidCookie.match(/sid=([^;]+)/);
+            if (sidMatch?.[1]) {
+              const sidToken = sidMatch[1];
+              return sidToken;
+            }
+          }
+        }
+      }
+
+      // if we can't establish a valid session this way, lets just warn the user and utilize the guest user context for the site
+      // eslint-disable-next-line no-console
+      console.warn(
+        `Warning: could not establish valid auth token for your site '${this.siteDisplayName}'.` +
+          'Local Dev proxied requests to your site may fail or return data from the guest user context.'
+      );
+
+      return ''; // Site will be guest user access only
+    }
+
+    // Not sure what scenarios we don't have an access token at all, but lets output a separate message here so we can distinguish these edge cases
+    // eslint-disable-next-line no-console
+    console.warn(
+      'Warning: sf cli org connection missing accessToken. Local Dev proxied requests to your site may fail or return data from the guest user context.'
+    );
+    return '';
   }
 }